requirements.txt
markdown2==2.4.10
pdfkit==1.0.0
openai==1.12.0
httpx==0.27.0
flask==2.3.3
werkzeug==2.3.7
flask-wtf==1.2.1
bs4==0.0.1
beautifulsoup4==4.12.2
requests==2.31.0
python-dateutil==2.8.2
python-dotenv==1.0.1
configparser==6.0.0
markdown==3.5.1 
selenium==4.15.2
pandas==2.1.3